requirements.txt
git+https://github.com/ZyMa-1/OsuFileParser.git@1.0.2
Pillow==10.0.0
PyYAML==6.0